+++ app/controllers/api/balance_summary_controller.rb
@@ -11,7 +11,9 @@ module Api
         return :not_in_allocation
       end
 
-      balance_summary = Rails.cache.fetch("balance_summary/#{allocation}", expires_in: 1.hours, race_condition_ttl: 3.seconds) do
+      cache_key = ["balance_summary", allocation, ::Configuration.gpu_account_pattern,
+                   ::Configuration.gpu_account_tres, ::Configuration.cpu_account_tres].join("/")
+      balance_summary = Rails.cache.fetch(cache_key, expires_in: 1.hours, race_condition_ttl: 3.seconds) do
         output = `scontrol show assoc accounts=#{allocation} flags=assoc -o | tail -n +3`
 
         if $?.success?
@@ -19,14 +21,14 @@ module Api
           account_line = Util.scontrol_to_hash(output).find { |line_h| line_h["Account"] == allocation }
           limit = if account_line
             grp_tres_mins = account_line["GrpTRESMins"].split(",").map { |pair| pair.split("=") }.to_h
-            data = account_line["Account"].end_with?("-gpu") ? grp_tres_mins["gres/gpu"] : grp_tres_mins["cpu"]
+            data = grp_tres_mins[::Configuration.account_tres_for(account_line["Account"])]
             data.match(/(\d+|N)\((\d+)\)/) { |m| m[1].to_f.positive? ? m[1].to_f / 60 : "No limit" }
           end
 
           # Then process the non-blank UserName lines as before, but use the found limit
           Util.scontrol_to_hash(output).select { |line_h| !line_h["UserName"].blank? }.map { |line_h|
             grp_tres_mins = line_h["GrpTRESMins"].split(",").map { |pair| pair.split("=") }.to_h
-            data = line_h["Account"].end_with?("-gpu") ? grp_tres_mins["gres/gpu"] : grp_tres_mins["cpu"]
+            data = grp_tres_mins[::Configuration.account_tres_for(line_h["Account"])]
             data.match(/(\d+|N)\((\d+)\)/) { |m| { 
               user: line_h["UserName"].split("(")[0], 
               used: m[2].to_f / 60,