+++ app/controllers/api/balance_usage_controller.rb
@@ -4,14 +4,19 @@ module Api
       user = @user.name
 
       allocations = Util.get_user_allocations(user)
-      mybalance = Rails.cache.fetch("balance_usage/#{user}", expires_in: 1.hours, race_condition_ttl: 3.seconds) do
+      # The TRES config participates in the key so that changing it takes effect
+      # immediately rather than after the hour-long cache expires.
+      cache_key = ["balance_usage", user, ::Configuration.gpu_account_pattern,
+                   ::Configuration.gpu_account_tres, ::Configuration.cpu_account_tres].join("/")
+      mybalance = Rails.cache.fetch(cache_key, expires_in: 1.hours, race_condition_ttl: 3.seconds) do
         output = `scontrol show assoc users=#{user} accounts=#{allocations} flags=assoc -o | tail -n +3`
 
         if $?.success?
           parsed_data = Util.scontrol_to_hash(output)
           parsed_data.select { |line_h| line_h["UserName"].blank? }.map { |line_h|
             grp_tres_mins = line_h["GrpTRESMins"].split(",").map { |pair| pair.split("=") }.to_h
-            data = line_h["Account"].end_with?("-gpu") ? grp_tres_mins["gres/gpu"] : grp_tres_mins["cpu"]
+            tres_key = ::Configuration.account_tres_for(line_h["Account"])
+            data = grp_tres_mins[tres_key]
             
             # Get user's specific usage for this account
             user_line = parsed_data.find { |l| 
@@ -19,7 +24,7 @@ module Api
               l["Account"] == line_h["Account"] 
             }
             user_tres_mins = user_line ? user_line["GrpTRESMins"].split(",").map { |pair| pair.split("=") }.to_h : {}
-            user_data = line_h["Account"].end_with?("-gpu") ? user_tres_mins["gres/gpu"] : user_tres_mins["cpu"]
+            user_data = user_tres_mins[tres_key]
             user_usage = user_data&.match(/(\d+|N)\((\d+)\)/)&.[](2).to_f / 60 rescue 0
             
             data.match(/(\d+|N)\((\d+)\)/) { |m|